// src/app/core/models/character_model.ts
export interface Character {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string | null;
    gender: string;
    origin: {
      name: string;
      url: string;
    };
    location: {
      name: string;
      url: string;
    };
    image: string;
    episode: string[];
    url: string;
    created: string;
  }
  
export interface CharacterInfo {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
    results: Character[];
}
  